Base Version
Base Version of the application, able to scan a code and pull out the
name encoded, look up into the Parse DB for that name and pull the
story and team of the person.

Using IntentIntegrator to handle the scan, not sure how it works for
apps without it installed.

Using same Parse code for pulling data.

Added a Insert button for any test data we would like to put in during
testing.

Took out commented code and put a big example of pulling and pushing
from Parse on the top, to keep the code clean and readable.