New Layout for InsertActivity
Changed InsertActivity layout to make it more flexible and fix
positioning of items, put them in a scroll view and on top of each
other so nothing clashes.

Chagned the Id's in the layout and Java file to follow naming
convention of View Id's being with Capital letters.

Got rid of text views and put hints inside EditTexts instead.