UI Changes, no more Insert, Comments, AsyncTask
UI is completely changed to a new design, only 2 screens right now.

Insert data was taken out since this functionality won't be in the app
for users, could be put into a seperate private app.

Commnets in the two updated Activities for documentation.

Tried to use a AsyncTask to help with the image loading, still freezing
the UI though. Not completely working yet.