Added LogIn page, first app open logic, donation tracking
Added LogIn page for users to log into Pullee application

Added logic so that if this is the first time the app is opened the
SignUp page is brought up, not LogIn.

Created new ParseObject in DB called Donation. This will be used to
keep track of who donates how much to who. When someone logged in the
app donates it enterst this information as a new Donation ParseObject.